refactor!: rename FilterResolver to QueryResolver
