+++ src/Exceptions/JsonApiException.php
@@ -23,7 +23,7 @@ class JsonApiException extends HttpException implements RenderableError
      * @param  string|null  $message
      * @param  Throwable|null  $previous
      * @param  int|null  $code
-     * @param  iterable|Error[]|null  $errors
+     * @param  iterable<Error>|null  $errors
      */
     public function __construct(
         int $statusCode = 500,
